pin appointment toggle (working)
