+++ src/modules/user/user.service.ts
@@ -8,7 +8,6 @@ import {PasswordReset} from "./password-reset/password-reset.entity";
 import {InvalidTokenException} from "../../exceptions/InvalidTokenException";
 import {DuplicateValueException} from "../../exceptions/DuplicateValueException";
 import {EmailChange} from "./email-change/email-change.entity";
-import {AppointmentService} from "../appointment/appointment.service";
 
 var crypto = require('crypto');
 var bcrypt = require('bcryptjs');
@@ -26,7 +25,6 @@ export class UserService {
         @InjectRepository(EmailChange)
         private readonly emailChangeRepository: Repository<EmailChange>,
         private mailerService: MailerService,
-        private appointmentService: AppointmentService
     ) {
     }
 
@@ -117,10 +115,7 @@ export class UserService {
     }
 
     public async findById(id: string): Promise<User | undefined> {
-        return await this.userRepository
-            .createQueryBuilder("user")
-            .where("user.id = :id", {id: id})
-            .getOne();
+        return await this.userRepository.findOne({where: {id: id}});
     }
 
     public async resetPasswordInit(mail: string, domain: string) {
@@ -425,19 +420,5 @@ export class UserService {
             "SET oldMail = 'invalid' " +
             "WHERE used IS NULL " +
             "AND userId = ?", [user.id]);
-    }
-
-    async pinAppointment(user: User, link: string) {
-        const appointment = await this.appointmentService.find(link);
-        const _user = await this.find(user.id);
-
-        if (_user.pinned.includes(appointment)) {
-            const removeIndex = _user.pinned.indexOf(appointment);
-            _user.pinned.splice(removeIndex, 1);
-        } else {
-            _user.pinned.push(appointment);
-        }
-
-        return this.userRepository.save(_user);
     }
 }
